changed clearTable
Deduplicated a lot of code, elements that would always be generated are now made at the beginning of the function.
The appending to ladderWindowTable is now put at the end of the function, as it would always be executed.
Created a small function to generate match buttons, to more clearly highlight the letter hidden inside it, and to save on duplicated code.

The usage of an array into a forEach is a stylistic choice, which also has the advantage of decluttering the appended object, this also makes it very slightly easier to add new objects to append.

Changed all instances of "let" to "const" to indicate that the value will never change.