changed updatePendingTable
Gave the different parts of data names, this makes the code far easier to read.

Changed the string concatenations to template literal substitution
Moved the column declarations to where they were first used, this makes it far easier to add/remove them.
made a template function for the column and button declarations, to make it easier to change them and add new ones later.

changed let to const where applicable.

changed the appending lines at the end to an array into forEach, making it easier to read the order.

TODO: make a construction that guarantees append order?